medium workable state with firebase request and active property
