+++ src/App.js
@@ -1,11 +1,4 @@
-// import { useState } from "react";
 import { useRouteMatch, Route, Switch, Redirect } from "react-router-dom";
-// import firebase from  "firebase";
-
-import firebase from "firebase/compat/app"; //firebase9!
-// import firebase from "firebase/compat/app"; //firebase9!
-import "firebase/compat/database"; //firebase9!
-// пробовали?
 
 import cn from 'classnames';
 import Footer from "./components/Footer";
@@ -18,35 +11,6 @@ import NotFoundPage from "./routes/NotFound";
 
 import s from './style.module.css';
 
-
-
-// Import the functions you need from the SDKs you need
-// import { initializeApp } from "firebase/app";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyAC-qjhzT6dPLAm355eE6dx2ZAA8bLX-O8",
-  authDomain: "react-marathon-pokemon-game.firebaseapp.com",
-  databaseURL: "https://react-marathon-pokemon-game-default-rtdb.firebaseio.com",
-  projectId: "react-marathon-pokemon-game",
-  storageBucket: "react-marathon-pokemon-game.appspot.com",
-  messagingSenderId: "55132643823",
-  appId: "1:55132643823:web:1122da5dbe19599bc7961a"
-};
-
-// Initialize Firebase
-// const app = initializeApp(firebaseConfig);
-
-firebase.initializeApp(firebaseConfig);
-const database = firebase.database();
-
-database.ref('pokemons').once('value', (snapshot) => {
-  console.log(`^_^ snapshot ↓ `);
-  // console.log(snapshot.val());
-  console.log(snapshot.val());
-});
 
 const App = () => {
