--- src/pokemons_data/pokemons_list.JSON
@@ -1,141 +1,0 @@
-{
-    "pokemons": {
-        "-MSjWsPjJfFkNKk6ixPF": {
-            "abilities": [
-                "keen-eye",
-                "tangled-feet",
-                "big-pecks"
-            ],
-            "base_experience": 122,
-            "height": 11,
-            "weight": 300,
-            "id": 17,
-            "img": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/17.png",
-            "name": "pidgeotto",
-            "stats": {
-                "hp": 63,
-                "attack": 60,
-                "defense": 55,
-                "special-attack": 50,
-                "special-defense": 50,
-                "speed": 71
-            },
-            "type": "normal",
-            "values": {
-                "top": 7,
-                "right": 5,
-                "bottom": 1,
-                "left": 2
-            }
-        },
-        "-MSjWsPo_nEu6XI67Jjz": {
-            "abilities": [
-                "intimidate",
-                "shed-skin",
-                "unnerve"
-            ],
-            "base_experience": 157,
-            "height": 35,
-            "weight": 650,
-            "id": 24,
-            "img": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/24.png",
-            "name": "arbok",
-            "stats": {
-                "hp": 60,
-                "attack": 95,
-                "defense": 69,
-                "special-attack": 65,
-                "special-defense": 79,
-                "speed": 80
-            },
-            "type": "poison",
-            "values": {
-                "top": 6,
-                "right": "A",
-                "bottom": 2,
-                "left": 8
-            }
-        },
-        "-MSjWsPo_nEu6XI67Jk-": {
-            "abilities": [
-                "static",
-                "lightning-rod"
-            ],
-            "base_experience": 112,
-            "height": 4,
-            "weight": 60,
-            "id": 25,
-            "img": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/25.png",
-            "name": "pikachu",
-            "stats": {
-                "hp": 35,
-                "attack": 55,
-                "defense": 40,
-                "special-attack": 50,
-                "special-defense": 50,
-                "speed": 90
-            },
-            "type": "electric",
-            "values": {
-                "top": 6,
-                "right": 5,
-                "bottom": 1,
-                "left": 5
-            }
-        },
-        "-MSjWsPps4Nua-6dEPjC": {
-            "abilities": [
-                "overgrow",
-                "chlorophyll"
-            ],
-            "base_experience": 64,
-            "height": 7,
-            "weight": 69,
-            "id": 1,
-            "img": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/1.png",
-            "name": "bulbasaur",
-            "stats": {
-                "hp": 45,
-                "attack": 49,
-                "defense": 49,
-                "special-attack": 65,
-                "special-defense": 65,
-                "speed": 45
-            },
-            "type": "grass",
-            "values": {
-                "top": 2,
-                "right": 2,
-                "bottom": 1,
-                "left": 4
-            }
-        },
-        "-MSjWsPq2RMOHZogNmW5": {
-            "abilities": [
-                "blaze",
-                "solar-power"
-            ],
-            "base_experience": 62,
-            "height": 6,
-            "weight": 85,
-            "id": 4,
-            "img": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/4.png",
-            "name": "charmander",
-            "stats": {
-                "hp": 39,
-                "attack": 52,
-                "defense": 43,
-                "special-attack": 60,
-                "special-defense": 50,
-                "speed": 65
-            },
-            "type": "fire",
-            "values": {
-                "top": 1,
-                "right": 1,
-                "bottom": 2,
-                "left": 1
-            }
-        }
-    }
-}
